[追加]接尾辞設定を反映
