+++ Niconicome/Models/Domain/Niconico/Download/Comment/CommentClient.cs
@@ -11,54 +11,9 @@ using Response = Niconicome.Models.Domain.Niconico.Net.Json.API.Comment.Response
 namespace Niconicome.Models.Domain.Niconico.Download.Comment
 {
 
-    public interface ICommentDownloadSettings
-    {
-        string NiconicoId { get; }
-        string FolderName { get; }
-        string FileNameFormat { get; }
-        bool IsOverwriteEnable { get; }
-        bool IsAutoDisposingEnable { get; }
-        bool IsDownloadingLogEnable { get; }
-        bool IsDownloadingOwnerCommentEnable { get; }
-        bool IsDownloadingEasyCommentEnable { get; }
-        bool IsReplaceStrictedEnable { get; }
-        bool IsUnsafeHandleEnable { get; }
-        int CommentOffset { get; }
-        int MaxcommentsCount { get; }
-    }
-
     public interface ICommentClient
     {
         Task<ICommentCollection> DownloadCommentAsync(IDmcInfo dmcInfo, ICommentDownloadSettings settings, IDownloadMessenger messenger, IDownloadContext context, CancellationToken token);
-    }
-
-    public class CommentDownloadSettings : ICommentDownloadSettings
-    {
-        public string NiconicoId { get; set; } = string.Empty;
-
-        public string FolderName { get; set; } = string.Empty;
-
-        public string FileNameFormat { get; set; } = string.Empty;
-
-        public bool IsOverwriteEnable { get; set; }
-
-        public bool IsAutoDisposingEnable { get; set; }
-
-        public bool IsDownloadingLogEnable { get; set; }
-
-        public bool IsDownloadingOwnerCommentEnable { get; set; }
-
-        public bool IsDownloadingEasyCommentEnable { get; set; }
-
-        public bool IsReplaceStrictedEnable { get; set; }
-
-        public bool IsUnsafeHandleEnable { get; set; }
-
-        public int CommentOffset { get; set; }
-
-        public int MaxcommentsCount { get; set; }
-
-
     }
 
     /// <summary>
